ts to js converted and backend integrated
