+++ .env
@@ -1,0 +1,1 @@
+VITE_API_URL=http://localhost:5000/api/admin
